WirelessNetwork: Fix reference access point for the active network
If the device is currently connected to a network, use the active
access point (the one it's actually connected to) as the reference
AP for this network.

This fixes the network parameters getting updates from an access
point that is not in use, even if it has stronger signal.

Also make sure we update the reference AP and signal strenght when
the active access point changes